refactor the bitmap function
